Fixed small issue

+++ professional-service/internal/domain/professional/model_professional.go
@@ -30,7 +30,7 @@ type Professional struct {
 
 type Availability struct {
 	ID             int64      `json:"-"`
-	ProfessionalID int        `json:"-"`
+	ProfessionalID int64      `json:"-"`
 	Date           civil.Date `json:"date,omitempty"`
 	From           civil.Time `json:"from,omitempty"`
 	To             civil.Time `json:"to,omitempty"`
